Revert onMetaData sending
That is an issue on Nginx-RTMP. The test showing there is no video stream
downloaded to the player through it with onMetaData function and being all
right without it confuses me quite a lot. Due to some users' complain, I
decide to revert onMetaData in repo and continue research in the future.

Signed-off-by: Leo Ma <begeekmyfriend@gmail.com>