Reduce streaming delay during connection
Yasea will not connect to the server until the first key frame is met.
But meanwhile the frame cache still accumulates some unnecessary frames
which would make serveral seconds delay on player end. Therefore we
change the strategy that the frame cache will not receive any frame
until the first key frame is met on connection.

Signed-off-by: Leo Ma <begeekmyfriend@gmail.com>