Optimize blur shader drawing
For external OES sampler, there is much overhead when drawing to the
window surface directly due to I/O sampling operation. So an optimized
approach is to draw the external OES sampler into FBO texture whose
resolution might be smaller than the window surface. And then we can
draw it as GLES internel sampler to the window surface with less
overhead.

There are two stages of drawing with a filter shader as external OES
sampler and another default shader as common 2D sampler.

Signed-off-by: Leo Ma <begeekmyfriend@gmail.com>