Fix white line on left side of the screen
The buffer address is forced to be aligned by allocateDirect method in
ByteBuffer so the valid data may not be stored from the first byte of the
buffer witch is still returned by the array method in ByteBuffer. So we need
to substitute allocate method for allocate direct method when fetching the
image from the camera.

Signed-off-by: begeekmyfriend <begeekmyfriend@gmail.com>